+++ frontend/user/response.py
@@ -1,0 +1,2 @@
+def generate_response(user_input):
+    return f"{user_input}"  # 그대로 출력
