+++ frontend/user/style.css
@@ -1,6 +1,17 @@
+@font-face {
+    font-family: 'Pretendard-Regular';
+    src: url('https://fastly.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Regular.woff') format('woff');
+    font-weight: 400;
+    font-style: normal;
+}
+
+:root {
+    --maincolor: yellow;
+}
+
 /* 기본 스타일 */
 body {
-    font-family: 'Arial', sans-serif;
+    font-family: 'Pretendard-Regular', sans-serif;
     background-color: #f7f7f7;
     padding: 20px;
 }
@@ -34,4 +45,5 @@ body {
     display: flex;
     flex-direction: column;
     align-items: flex-start;
+    background-color: var(--maincolor);
 }
